fix indirect_x

+++ src/cpu.jl
@@ -280,9 +280,9 @@ function address(cpu::CPU, mode::AddressingMode, logger::StepLogger)::Tuple{UInt
         addr = params2addr(lo, hi) + cpu.register_y
     elseif mode == indirect_x
         base = lo
-        ptr = UInt16(base + cpu.register_x)
-        lo = read8(cpu.bus, ptr)
-        hi = read8(cpu.bus, ptr + 0x01)
+        ptr = base + cpu.register_x
+        lo = read8(cpu.bus, UInt16(ptr))
+        hi = read8(cpu.bus, UInt16(ptr + 0x01))
         addr = params2addr(lo, hi)
     elseif mode == indirect_y
         base = lo
